CLOUD-45782 storing images in versioned component table
