+++ cloud-arm/build.gradle
@@ -33,6 +33,7 @@ dependencies {
 
     testCompile project(path: ':core-model', configuration: 'tests')
     testCompile group: 'org.mockito',               name: 'mockito-all',                    version: '1.10.19'
+    testCompile group: 'org.springframework.boot',  name: 'spring-boot-starter',            version: springBootVersion
     testCompile group: 'org.springframework.boot',  name: 'spring-boot-starter-test',       version: '1.1.8.RELEASE'
     testCompile group: 'junit',                     name: 'junit',                          version: '4.12'
