+++ cloud-common/build.gradle
@@ -13,6 +13,8 @@ jar {
 
 dependencies {
   compile group: 'org.slf4j',                     name: 'slf4j-api',                      version: slf4jApiVersion
+  // TODO remove spring boot dependency
   compile group: 'org.springframework.boot',      name: 'spring-boot-starter',            version: springBootVersion
   compile group: 'net.sf.ehcache',                name: 'ehcache',                        version: '2.8.3'
+  compile group: 'com.fasterxml.jackson.core',    name: 'jackson-databind',               version: jacksonVersion
 }
