+++ cloud-common/src/main/java/com/sequenceiq/cloudbreak/common/type/ComponentType.java
@@ -1,0 +1,5 @@
+package com.sequenceiq.cloudbreak.common.type;
+
+public enum ComponentType {
+    IMAGE, CONTAINER
+}
