+++ cloud-reactor/build.gradle
@@ -26,6 +26,7 @@ dependencies {
 
     testCompile group: 'org.mockito',               name: 'mockito-all',                    version: '1.10.19'
     testCompile group: 'junit',                     name: 'junit',                          version: '4.12'
+    testCompile group: 'org.springframework.boot',  name: 'spring-boot-starter',            version: springBootVersion
     testCompile group: 'org.springframework.boot',  name: 'spring-boot-starter-test',       version: '1.1.8.RELEASE'
 
 }
