+++ cloud-reactor/src/main/java/com/sequenceiq/cloudbreak/cloud/service/CloudbreakResourceNameService.java
