--- core-model/src/main/java/com/sequenceiq/cloudbreak/domain/StackDescription.java
@@ -1,5 +1,0 @@
-package com.sequenceiq.cloudbreak.domain;
-
-public abstract class StackDescription {
-
-}
