+++ core-model/src/main/java/com/sequenceiq/cloudbreak/domain/json/JsonToString.java
@@ -1,0 +1,18 @@
+package com.sequenceiq.cloudbreak.domain.json;
+
+import javax.persistence.AttributeConverter;
+
+public class JsonToString implements AttributeConverter<Json, String> {
+    @Override
+    public String convertToDatabaseColumn(Json attribute) {
+        if (attribute != null) {
+            return attribute.getValue();
+        }
+        return null;
+    }
+
+    @Override
+    public Json convertToEntityAttribute(String dbData) {
+        return new Json(dbData);
+    }
+}
