+++ core/src/main/java/com/sequenceiq/cloudbreak/repository/ComponentRepository.java
@@ -1,0 +1,16 @@
+package com.sequenceiq.cloudbreak.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.query.Param;
+
+import com.sequenceiq.cloudbreak.common.type.ComponentType;
+import com.sequenceiq.cloudbreak.domain.Component;
+
+@EntityType(entityClass = Component.class)
+public interface ComponentRepository extends CrudRepository<Component, Long> {
+
+    Component findComponentByStackIdComponentTypeName(@Param("stackId") Long stackId, @Param("componentType") ComponentType componentType,
+            @Param("name") String name);
+
+
+}
