+++ core/src/main/java/com/sequenceiq/cloudbreak/service/cluster/flow/InstanceTerminationHandler.java
@@ -2,7 +2,6 @@ package com.sequenceiq.cloudbreak.service.cluster.flow;
 
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
 import javax.inject.Inject;
@@ -11,27 +10,21 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
-import com.sequenceiq.cloudbreak.core.CloudbreakSecuritySetupException;
-import com.sequenceiq.cloudbreak.common.type.CloudPlatform;
-import com.sequenceiq.cloudbreak.domain.InstanceGroup;
-import com.sequenceiq.cloudbreak.common.type.InstanceGroupType;
-import com.sequenceiq.cloudbreak.domain.InstanceMetaData;
 import com.sequenceiq.cloudbreak.common.type.InstanceStatus;
+import com.sequenceiq.cloudbreak.common.type.Status;
+import com.sequenceiq.cloudbreak.domain.InstanceGroup;
+import com.sequenceiq.cloudbreak.domain.InstanceMetaData;
 import com.sequenceiq.cloudbreak.domain.Resource;
 import com.sequenceiq.cloudbreak.domain.Stack;
-import com.sequenceiq.cloudbreak.common.type.Status;
 import com.sequenceiq.cloudbreak.repository.HostMetadataRepository;
 import com.sequenceiq.cloudbreak.repository.InstanceGroupRepository;
 import com.sequenceiq.cloudbreak.repository.InstanceMetaDataRepository;
 import com.sequenceiq.cloudbreak.repository.ResourceRepository;
 import com.sequenceiq.cloudbreak.service.CloudPlatformResolver;
-import com.sequenceiq.cloudbreak.service.CloudbreakServiceException;
-import com.sequenceiq.cloudbreak.service.TlsSecurityService;
 import com.sequenceiq.cloudbreak.service.events.CloudbreakEventService;
 import com.sequenceiq.cloudbreak.service.messages.CloudbreakMessagesService;
 import com.sequenceiq.cloudbreak.service.stack.connector.CloudPlatformConnector;
 import com.sequenceiq.cloudbreak.service.stack.connector.MetadataSetup;
-import com.sequenceiq.cloudbreak.service.stack.connector.UserDataBuilder;
 
 @Service
 public class InstanceTerminationHandler {
@@ -52,10 +45,7 @@ public class InstanceTerminationHandler {
     private CloudPlatformResolver platformResolver;
     @Inject
     private CloudbreakMessagesService cloudbreakMessagesService;
-    @Inject
-    private UserDataBuilder userDataBuilder;
-    @Inject
-    private TlsSecurityService tlsSecurityService;
+
 
     private enum Msg {
         STACK_INSTANCE_TERMINATE("stack.instance.terminate"),
@@ -97,19 +87,8 @@ public class InstanceTerminationHandler {
         CloudPlatformConnector cloudPlatformConnector = platformResolver.connector(stack.cloudPlatform());
         Set<String> instanceIds = new HashSet<>();
         instanceIds.add(instanceMetaData.getInstanceId());
-        Map<InstanceGroupType, String> userdata = buildUserData(stack, stack.cloudPlatform(), cloudPlatformConnector);
-        cloudPlatformConnector.removeInstances(stack, userdata.get(InstanceGroupType.GATEWAY),
-                userdata.get(InstanceGroupType.CORE), instanceIds, instanceMetaData.getInstanceGroup().getGroupName());
+        cloudPlatformConnector.removeInstances(stack, instanceIds, instanceMetaData.getInstanceGroup().getGroupName());
         LOGGER.info("Instance [ id: {}, name: {} ] deleted", instanceMetaData.getId(), instanceMetaData.getInstanceId());
-    }
-
-    private Map<InstanceGroupType, String> buildUserData(Stack stack, CloudPlatform platform, CloudPlatformConnector connector) {
-        try {
-            return userDataBuilder.buildUserData(platform, tlsSecurityService.readPublicSshKey(stack.getId()), stack.getCredential().getLoginUserName(),
-                    connector.getPlatformParameters(stack));
-        } catch (CloudbreakSecuritySetupException e) {
-            throw new CloudbreakServiceException(e.getMessage(), e);
-        }
     }
 
     private void deleteInstanceResourceFromDatabase(Stack stack, InstanceMetaData instanceMetaData) {
