+++ core/src/main/java/com/sequenceiq/cloudbreak/service/stack/connector/ProvisionSetup.java
@@ -13,14 +13,6 @@ public interface ProvisionSetup {
 
     ImageStatusResult checkImage(Stack stack) throws Exception;
 
-    /**
-     * @deprecated There is no pre-provision check in SPI, therefore this method will be deleted soon
-     *
-     *
-     */
-    @Deprecated
-    String preProvisionCheck(Stack stack);
-
     CloudPlatform getCloudPlatform();
 
 }
