+++ core/src/main/java/com/sequenceiq/cloudbreak/service/stack/flow/ProvisioningService.java
@@ -9,16 +9,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
-import com.sequenceiq.cloudbreak.core.CloudbreakSecuritySetupException;
 import com.sequenceiq.cloudbreak.common.type.CloudPlatform;
-import com.sequenceiq.cloudbreak.common.type.InstanceGroupType;
+import com.sequenceiq.cloudbreak.core.CloudbreakSecuritySetupException;
 import com.sequenceiq.cloudbreak.domain.Resource;
 import com.sequenceiq.cloudbreak.domain.Stack;
 import com.sequenceiq.cloudbreak.repository.StackRepository;
 import com.sequenceiq.cloudbreak.service.CloudPlatformResolver;
 import com.sequenceiq.cloudbreak.service.TlsSecurityService;
 import com.sequenceiq.cloudbreak.service.stack.connector.CloudPlatformConnector;
-import com.sequenceiq.cloudbreak.service.stack.connector.UserDataBuilder;
 import com.sequenceiq.cloudbreak.service.stack.event.ProvisionComplete;
 
 @Service
@@ -33,18 +31,13 @@ public class ProvisioningService {
     private CloudPlatformResolver cloudPlatformResolver;
 
     @Inject
-    private UserDataBuilder userDataBuilder;
-
-    @Inject
     private TlsSecurityService tlsSecurityService;
 
     public ProvisionComplete buildStack(final CloudPlatform cloudPlatform, Stack stack, Map<String, Object> setupProperties)
             throws CloudbreakSecuritySetupException {
         stack = stackRepository.findOneWithLists(stack.getId());
         CloudPlatformConnector connector = cloudPlatformResolver.connector(cloudPlatform);
-        Map<InstanceGroupType, String> userdata = userDataBuilder.buildUserData(cloudPlatform, tlsSecurityService.readPublicSshKey(stack.getId()),
-                stack.getCredential().getLoginUserName(), connector.getPlatformParameters(stack));
-        Set<Resource> resources = connector.buildStack(stack, userdata.get(InstanceGroupType.GATEWAY), userdata.get(InstanceGroupType.CORE), setupProperties);
+        Set<Resource> resources = connector.buildStack(stack, setupProperties);
         return new ProvisionComplete(cloudPlatform, stack.getId(), resources);
     }
 }
