+++ core/src/test/java/com/sequenceiq/cloudbreak/core/bootstrap/service/ClusterBootstrapperErrorHandlerTest.java
@@ -26,10 +26,10 @@ import org.mockito.stubbing.Answer;
 
 import com.sequenceiq.cloudbreak.TestUtil;
 import com.sequenceiq.cloudbreak.common.type.CloudPlatform;
+import com.sequenceiq.cloudbreak.common.type.ResourceType;
 import com.sequenceiq.cloudbreak.domain.InstanceGroup;
 import com.sequenceiq.cloudbreak.domain.InstanceMetaData;
 import com.sequenceiq.cloudbreak.domain.Resource;
-import com.sequenceiq.cloudbreak.common.type.ResourceType;
 import com.sequenceiq.cloudbreak.domain.Stack;
 import com.sequenceiq.cloudbreak.orchestrator.ContainerOrchestrator;
 import com.sequenceiq.cloudbreak.orchestrator.exception.CloudbreakOrchestratorFailedException;
@@ -40,12 +40,10 @@ import com.sequenceiq.cloudbreak.repository.InstanceGroupRepository;
 import com.sequenceiq.cloudbreak.repository.InstanceMetaDataRepository;
 import com.sequenceiq.cloudbreak.repository.ResourceRepository;
 import com.sequenceiq.cloudbreak.service.CloudPlatformResolver;
-import com.sequenceiq.cloudbreak.service.TlsSecurityService;
 import com.sequenceiq.cloudbreak.service.events.CloudbreakEventService;
 import com.sequenceiq.cloudbreak.service.messages.CloudbreakMessagesService;
 import com.sequenceiq.cloudbreak.service.stack.connector.CloudPlatformConnector;
 import com.sequenceiq.cloudbreak.service.stack.connector.MetadataSetup;
-import com.sequenceiq.cloudbreak.service.stack.connector.UserDataBuilder;
 import com.sequenceiq.cloudbreak.service.stack.resource.ResourceBuilder;
 import com.sequenceiq.cloudbreak.service.stack.resource.ResourceBuilderInit;
 
@@ -90,12 +88,6 @@ public class ClusterBootstrapperErrorHandlerTest {
 
     @Mock
     private CloudbreakMessagesService cloudbreakMessagesService;
-
-    @Mock
-    private UserDataBuilder userDataBuilder;
-
-    @Mock
-    private TlsSecurityService tlsSecurityService;
 
     @InjectMocks
     private ClusterBootstrapperErrorHandler underTest;
@@ -152,7 +144,7 @@ public class ClusterBootstrapperErrorHandlerTest {
         when(metadataSetup.getInstanceResourceType()).thenReturn(ResourceType.AZURE_VIRTUAL_MACHINE);
         doNothing().when(resourceRepository).delete(anyLong());
         when(resourceRepository.findByStackIdAndNameAndType(anyLong(), anyString(), any(ResourceType.class))).thenReturn(new Resource());
-        when(cloudPlatformConnector.removeInstances(any(Stack.class), anyString(), anyString(), anySet(), anyString())).thenReturn(new HashSet<String>());
+        when(cloudPlatformConnector.removeInstances(any(Stack.class), anySet(), anyString())).thenReturn(new HashSet<String>());
         when(instanceMetaDataRepository.findNotTerminatedByPrivateAddress(anyLong(), anyString())).thenAnswer(new Answer<InstanceMetaData>() {
             @Override
             public InstanceMetaData answer(InvocationOnMock invocation) {
@@ -187,7 +179,7 @@ public class ClusterBootstrapperErrorHandlerTest {
         verify(instanceGroupRepository, times(3)).save(any(InstanceGroup.class));
         verify(instanceMetaDataRepository, times(3)).save(any(InstanceMetaData.class));
         verify(platformResolver, times(3)).connector(any(CloudPlatform.class));
-        verify(cloudPlatformConnector, times(3)).removeInstances(any(Stack.class), anyString(), anyString(), anySet(), anyString());
+        verify(cloudPlatformConnector, times(3)).removeInstances(any(Stack.class), anySet(), anyString());
         verify(metadataSetup, times(3)).getInstanceResourceType();
         verify(resourceRepository, times(3)).findByStackIdAndNameAndType(anyLong(), anyString(), any(ResourceType.class));
         verify(platformResolver, times(3)).metadata(any(CloudPlatform.class));
