+++ gradle.properties
@@ -7,3 +7,4 @@ eventBusVersion=2.0.3.RELEASE
 slf4jApiVersion=1.7.12
 guavaVersion=16.0.1
 freemarkerVersion=2.3.20
+jacksonVersion=2.6.0
